Limit parsing to line-only.
The temporary string buffer was the string source for the complete
paragraph, preventing us from using it because the borrow was not
from raw_lines (the 'a lifetime). This could be fixed by either
transforming buffer (local) borrows to raw_lines, or somehow
creating a &'a str view which pest is able to iterate over.