Move object.rs into mod.rs
The "object.rs" pattern was to prevent mod.rs files from being too large.
Instead it creates an unnecessary file, so we just join it in since the
mod.rs files aren't that big anyways.

Additionally what "object.rs" means isn't as clear as just keeping it in
mod.rs