Remove get_index_end().
We don't need it at all, because slices are [x, y),
the start of a slice is always the end of a previous one.

This means it's in the map, and correctly points to where
we want to go. Easy.

I'm honestly not sure how the implementation prior to this
worked at all now that I think about it.