Wrap OwnedElementsIterator::Single in a Box.
This address a clippy lint involving a large difference in sizes
between enum variants.

However, because this adds an allocation (and an empty Vec is free),
there is now no longer any purpose in having separate enum cases.
The plan is now to replace the OwnedElementsIterator implementation
with a Vec<Element<'t>>.